+++ src/main/java/nl/nn/testtool/echo2/reports/PathComponent.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2018 Nationale-Nederlanden
+   Copyright 2018 Nationale-Nederlanden, 2020 WeAreFrank!
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -26,10 +26,7 @@ import nl.nn.testtool.echo2.Echo2Application;
 import echopointng.tree.DefaultMutableTreeNode;
 
 /**
- * @author m00f069
- *
- * To change the template for this generated type comment go to
- * Window&gt;Preferences&gt;Java&gt;Code Generation&gt;Code and Comments
+ * @author Jaco de Groot
  */
 public class PathComponent extends Column implements ActionListener {
 	private TreePane treePane;