+++ src/main/java/nl/nn/testtool/echo2/reports/ReportsTreeCellRenderer.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2018 Nationale-Nederlanden
+   Copyright 2018 Nationale-Nederlanden, 2020 WeAreFrank!
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -15,7 +15,10 @@
 */
 package nl.nn.testtool.echo2.reports;
 
-import org.apache.log4j.Logger;
+import java.lang.invoke.MethodHandles;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import echopointng.Tree;
 import echopointng.tree.DefaultMutableTreeNode;
@@ -27,16 +30,12 @@ import nextapp.echo2.app.ResourceImageReference;
 import nl.nn.testtool.Checkpoint;
 import nl.nn.testtool.Report;
 import nl.nn.testtool.echo2.Echo2Application;
-import nl.nn.testtool.util.LogUtil;
 
 /**
- * @author m00f069
- *
- * To change the template for this generated type comment go to
- * Window&gt;Preferences&gt;Java&gt;Code Generation&gt;Code and Comments
+ * @author Jaco de Groot
  */
 public class ReportsTreeCellRenderer extends DefaultTreeCellRenderer {
-	private Logger log = LogUtil.getLogger(this);
+	private Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 	private boolean showReportAndCheckpointIds;
 
 	/**