+++ src/main/java/nl/nn/testtool/storage/file/TestStorage.java
@@ -27,19 +27,16 @@ import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 
+import org.slf4j.Logger;
+
 import nl.nn.testtool.MetadataExtractor;
 import nl.nn.testtool.Report;
 import nl.nn.testtool.storage.StorageException;
 import nl.nn.testtool.util.CSVReader;
 import nl.nn.testtool.util.SearchUtil;
 
-import org.apache.log4j.Logger;
-
 /**
- * @author m00f069
- *
- * To change the template for this generated type comment go to
- * Window&gt;Preferences&gt;Java&gt;Code Generation&gt;Code and Comments
+ * @author Jaco de Groot
  */
 public class TestStorage implements nl.nn.testtool.storage.CrudStorage {
 //TODO TestStorage -> CrudStorage?